+++ Source/Examples/Xamarin.Android/ExampleBrowser/CategoryListActivity.cs
@@ -4,15 +4,15 @@
 // </copyright>
 // --------------------------------------------------------------------------------------------------------------------
 
-using System.Linq;
-
-using Android.App;
-using Android.Content;
-using Android.OS;
-using Android.Widget;
-
 namespace ExampleBrowser
 {
+    using System.Linq;
+
+    using Android.App;
+    using Android.Content;
+    using Android.OS;
+    using Android.Widget;
+
     [Activity(Label = "OxyPlot Example Browser", MainLauncher = true, Icon = "@drawable/icon")]
     public class CategoryListActivity : ListActivity
     {
@@ -21,16 +21,16 @@ namespace ExampleBrowser
             base.OnCreate(bundle);
             this.Title = "OxyPlot Example Browser";
 
-            var examples=ExampleLibrary.Examples.GetList();
+            var examples = ExampleLibrary.Examples.GetList();
             var categories = examples.Select(e => e.Category).Distinct().OrderBy(s => s).ToList();
-            ListAdapter = new ArrayAdapter<string>(this, Resource.Layout.ListItem, categories);
-            ListView.TextFilterEnabled = true;
-            ListView.ItemClick += delegate(object sender, AdapterView.ItemClickEventArgs args)
+            this.ListAdapter = new ArrayAdapter<string>(this, Resource.Layout.ListItem, categories);
+            this.ListView.TextFilterEnabled = true;
+            this.ListView.ItemClick += delegate (object sender, AdapterView.ItemClickEventArgs args)
             {
                 var category = categories[args.Position];
                 var second = new Intent(this, typeof(ExampleListActivity));
                 second.PutExtra("category", category);
-                StartActivity(second);
+                this.StartActivity(second);
             };
         }
     }