+++ Source/Examples/Xamarin.Android/ExampleBrowser/ExampleListActivity.cs
@@ -4,37 +4,37 @@
 // </copyright>
 // --------------------------------------------------------------------------------------------------------------------
 
-using System.Linq;
-
-using Android.App;
-using Android.Content;
-using Android.OS;
-using Android.Widget;
-
 namespace ExampleBrowser
 {
+    using System.Linq;
+
+    using Android.App;
+    using Android.Content;
+    using Android.OS;
+    using Android.Widget;
+
     [Activity(Label = "Example list", Icon = "@drawable/icon")]
     public class ExampleListActivity : ListActivity
     {
         protected override void OnCreate(Bundle bundle)
         {
             base.OnCreate(bundle);
-            var category = Intent.GetStringExtra("category");
+            var category = this.Intent.GetStringExtra("category");
             this.Title = category;
 
             var examples = ExampleLibrary.Examples.GetList();
             var plots = examples.Where(e => e.Category == category).Select(e => e.Title).OrderBy(s => s).ToList();
-            ListAdapter = new ArrayAdapter<string>(this, Resource.Layout.ListItem, plots);
-            ListView.TextFilterEnabled = true;
-            ListView.ItemClick += delegate(object sender, AdapterView.ItemClickEventArgs args)
+            this.ListAdapter = new ArrayAdapter<string>(this, Resource.Layout.ListItem, plots);
+            this.ListView.TextFilterEnabled = true;
+            this.ListView.ItemClick += delegate (object sender, AdapterView.ItemClickEventArgs args)
             {
                 // When clicked, show a toast with the TextView text
-                Toast.MakeText(Application, ((TextView)args.View).Text, ToastLength.Short).Show();
+                Toast.MakeText(this.Application, ((TextView)args.View).Text, ToastLength.Short).Show();
                 var second = new Intent(this, typeof(PlotActivity));
 
                 second.PutExtra("category", category);
                 second.PutExtra("plot", plots[args.Position]);
-                StartActivity(second);
+                this.StartActivity(second);
             };
         }
     }