+++ Source/Examples/Xamarin.Mac/SimpleDemo/Main.cs
@@ -1,7 +1,7 @@
-﻿using AppKit;
-
 namespace SimpleDemo
 {
+    using AppKit;
+
     static class MainClass
     {
         static void Main(string[] args)